+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Added `Lightness` mutation to make colors darker / lighter.
+- Added `Color Transition` mutation to transition color to target color.
+- Added `Target Color` mutation as target for Colorize.
+- Added option to reload texture from disk.
+
 ## [1.2.0] - 2021-05-05
 
 ### Added